ml-pipeline/requirements.txt
# ML Pipeline Dependencies
# Python >= 3.11 required

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Redis Streams consumer
redis[asyncio]==5.2.1

# Data engineering
pandas==2.2.3
numpy==2.1.3

# ML training
scikit-learn==1.5.2
joblib==1.4.2

# Data validation
pydantic==2.9.2

# Utilities
python-dotenv==1.0.1
